refactor: delete prev DicomViewer component
