+++ apps/insight-viewer/src/app/ControlledViewer.tsx
@@ -1,4 +1,4 @@
-import { useDicomViewer, NewDicomViewer } from '@lunit-insight-viewer/react';
+import { useDicomViewer, DicomViewer } from '@lunit-insight-viewer/react';
 
 import { imageIds, tools } from './image';
 
@@ -37,7 +37,7 @@ export function ControlledViewer() {
 
   return (
     <div style={{ width: '500px', height: '500px' }}>
-      <NewDicomViewer ref={viewerRef} />
+      <DicomViewer ref={viewerRef} />
       <button onClick={handleRotateButtonClick}>Rotate 30</button>
       <div style={{ marginBottom: '8px' }}>
         {viewerInfo?.viewport.properties.rotation}
