+++ libs/react-dicom-viewer/src/index.ts
@@ -1,3 +1,2 @@
 export * from './DicomViewer';
-export * from './NewDicomViewer';
 export * from './useDicomViewer';
